Production/victim/requirements.txt
pycryptodome>=3.15.0
dnslib>=0.9.7